♻️ refactor: application dto 를 domain 계층의 vo 로 변경
infra 에서 application 의존성을 없애기 위함
